+++ mountpoint-s3-fs/tests/common/mod.rs
@@ -12,6 +12,9 @@ pub mod fuse;
 #[cfg(feature = "s3_tests")]
 pub mod s3;
 
+#[cfg(feature = "manifest")]
+pub mod manifest;
+
 use aws_credential_types::Credentials;
 use fuser::{FileAttr, FileType};
 use futures::executor::ThreadPool;
