+++ mountpoint-s3-fs/tests/fuse_tests/mod.rs
@@ -2,8 +2,6 @@
 mod cache_test;
 mod consistency_test;
 mod lookup_test;
-#[cfg(feature = "manifest")]
-mod manifest_test;
 mod mkdir_test;
 mod perm_test;
 mod prefetch_test;
