+++ mountpoint-s3-fs/tests/mod.rs
@@ -2,4 +2,6 @@ pub mod common;
 
 #[cfg(feature = "fuse_tests")]
 mod fuse_tests;
+#[cfg(feature = "manifest")]
+mod manifest_tests;
 mod reftests;
